+++ macro/src/lib.rs
@@ -62,7 +62,9 @@ pub fn derive_mmio(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
         .iter()
         .map(|field| (field, field.ident.as_ref().unwrap()))
         .filter(|(_field, field_ident)| !field_ident.to_string().starts_with("_"))
-        .map(|(field, field_ident)| field_parser.generate_access_methods(field, field_ident));
+        .map(|(field, field_ident)| {
+            field_parser.generate_access_methods(&ident, field, field_ident)
+        });
 
     let access_methods_quoted = quote! {
         #(#access_methods)*
@@ -83,25 +85,32 @@ pub fn derive_mmio(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
         None
     } else {
         Some(quote! {
-            #[doc = "Create a new handle to this peripheral given an address."]
-            #[doc = ""]
-            #[doc = "# Safety"]
-            #[doc = ""]
-            #[doc = "See the safety notes for [`new_mmio`]."]
+            /// Create a new handle to this peripheral given an address.
+            ///
+            /// # Safety
+            ///
+            /// See the safety notes for [Self::new_mmio].
             pub const unsafe fn new_mmio_at(addr: usize) -> #wrapper_ident<'static> {
                 Self::new_mmio(addr as *mut #ident)
             }
 
-            #[doc = "Create a new handle to this peripheral."]
-            #[doc = ""]
-            #[doc = "# Safety"]
-            #[doc = ""]
-            #[doc = "The pointer given must have suitable alignment, and point to an object"]
-            #[doc = "which matches the layout given by the structure pointed to."]
-            #[doc = ""]
-            #[doc = "If you create multiple instances of this handle at the same time,"]
-            #[doc = "you are responsible for ensuring that there are no read-modify-write"]
-            #[doc = "races on any of the registers."]
+            /// Create a new handle to this peripheral.
+            ///
+            /// # Safety
+            ///
+            /// The pointer given must have suitable alignment, and point to an object
+            /// which matches the layout given by the structure pointed to.
+            ///
+            /// If you create multiple instances of this handle at the same time,
+            /// you are responsible for ensuring that there are no read-modify-write
+            /// races on any of the registers.
+            ///
+            /// The [core::marker::Send] is implemented for the MMIO block.
+            /// However, there are cases where this implementation might be incorrect, for example
+            /// if an MMIO block was created for a core-local private address.
+            ///
+            /// In that case, it it is recommended to wrap the MMIO block in a newtype structure
+            /// and [un-implement Send](https://doc.rust-lang.org/nomicon/send-and-sync.html).
             #[inline]
             pub const unsafe fn new_mmio(ptr: *mut #ident) -> #wrapper_ident<'static> {
                 #wrapper_ident {
@@ -114,9 +123,9 @@ pub fn derive_mmio(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
 
     // combine the fragments into the desired output code
     proc_macro::TokenStream::from(quote! {
-        #[doc = "An MMIO wrapper for [`"]
+        #[doc = "An MMIO wrapper for ["]
         #[doc = stringify!(#ident)]
-        #[doc = "`]"]
+        #[doc = "]"]
         pub struct #wrapper_ident<'a> {
             ptr: *mut #ident,
             phantom: core::marker::PhantomData<&'a ()>,
@@ -149,6 +158,16 @@ pub fn derive_mmio(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
 
         unsafe impl derive_mmio::_MmioMarker for #wrapper_ident<'_> {}
 
+        /// Send is unsafely implemented because sending a register block pointer to another
+        /// thread should not be an issue for most use-cases.
+        ///
+        /// However, there are cases where this [core::marker::Send] implementation might be
+        /// invalid, for example if an MMIO block was created for a core-local private address.
+        ///
+        /// In that case, it it is recommended to wrap the MMIO block in a newtype structure
+        /// and [un-implement Send](https://doc.rust-lang.org/nomicon/send-and-sync.html).
+        unsafe impl core::marker::Send for #wrapper_ident<'_> {}
+
         impl #ident {
             #bound_check_func
 
@@ -179,7 +198,12 @@ struct FieldParser {
 
 impl FieldParser {
     /// Convert a field into a set of methods that operate on that field
-    fn generate_access_methods(&mut self, field: &Field, field_ident: &Ident) -> TokenStream {
+    fn generate_access_methods(
+        &mut self,
+        ident: &Ident,
+        field: &Field,
+        field_ident: &Ident,
+    ) -> TokenStream {
         let mut access = None;
         for attr in field.attrs.iter() {
             if attr.path().is_ident("mmio") {
@@ -218,10 +242,22 @@ impl FieldParser {
         let mut output = TokenStream::new();
         match &field.ty {
             syn::Type::Array(type_array) => {
-                self.generate_array_access_methods(access, field_ident, type_array, &mut output);
+                self.generate_array_access_methods(
+                    ident,
+                    access,
+                    field_ident,
+                    type_array,
+                    &mut output,
+                );
             }
             syn::Type::Path(type_path) => {
-                self.generate_field_access_methods(access, field_ident, type_path, &mut output);
+                self.generate_field_access_methods(
+                    ident,
+                    access,
+                    field_ident,
+                    type_path,
+                    &mut output,
+                );
             }
             _ => (),
         }
@@ -256,18 +292,22 @@ impl FieldParser {
                     segments,
                     leading_colon: type_path.path.leading_colon,
                 };
+                let inner_mmio_path_doc_ref = inner_mmio_path
+                    .to_token_stream()
+                    .to_string()
+                    .replace(" :: ", "::");
                 self.bound_checks.push(quote! {
                     derive_mmio::is_mmio::<#inner_mmio_path>();
                 });
                 let steal_func_name = format_ident!("steal_{}", field_ident);
                 quote! {
-                    #[doc = "Obtain a reference to the inner MMIO field `"]
-                    #[doc = stringify!(#field_ident)]
-                    #[doc = "`."]
-                    #[doc = "# Lifetime"]
+                    #[doc = "Obtain a reference to the inner MMIO field "]
+                    #[doc = concat!("[", #inner_mmio_path_doc_ref, "]")]
                     #[doc = ""]
-                    #[doc = "The lifetime of the returned inner MMIO block is tied to the "]
-                    #[doc = "lifetime of this structure"]
+                    /// # Lifetime
+                    ///
+                    /// The lifetime of the returned inner MMIO block is tied to the
+                    /// lifetime of this structure
                     #[inline]
                     pub fn #field_ident(&mut self) -> #inner_mmio_path<'_> {
                         unsafe {
@@ -275,17 +315,17 @@ impl FieldParser {
                         }
                     }
 
-                    #[doc = "Obtain a reference to the inner MMIO field `"]
-                    #[doc = stringify!(#field_ident)]
-                    #[doc = "`."]
-                    #[doc = "# Lifetime and Safety"]
+                    #[doc = "Unsafely obtain a reference to the inner MMIO field `"]
+                    #[doc = concat!("[", #inner_mmio_path_doc_ref, "]")]
                     #[doc = ""]
-                    #[doc = "The lifetime of the returned inner MMIO block is static which"]
-                    #[doc = "allows independent usage of the inner block and arbitrary"]
-                    #[doc = "creation of of multiple inner blocks for the same peripheral."]
-                    #[doc = "If you create multiple instances of this handle at the same time,"]
-                    #[doc = "you are responsible for ensuring that there are no read-modify-write"]
-                    #[doc = "races on any of the registers."]
+                    /// # Lifetime and Safety
+                    ///
+                    /// The lifetime of the returned inner MMIO block is static which
+                    /// allows independent usage of the inner block and arbitrary
+                    /// creation of of multiple inner blocks for the same peripheral.
+                    /// If you create multiple instances of this handle at the same time,
+                    /// you are responsible for ensuring that there are no read-modify-write
+                    /// races on any of the registers.
                     #[inline]
                     pub unsafe fn #steal_func_name(&mut self) -> #inner_mmio_path<'static> {
                         let ptr = unsafe { core::ptr::addr_of_mut!((*self.ptr).#field_ident) };
@@ -305,6 +345,7 @@ impl FieldParser {
     }
     fn generate_field_access_methods(
         &self,
+        ident: &Ident,
         access: Access,
         field_ident: &Ident,
         type_path: &TypePath,
@@ -316,9 +357,9 @@ impl FieldParser {
         let modify_fn_name = format_ident!("modify_{}", field_ident);
 
         access_methods.append_all(quote! {
-            #[doc = "Obtain a pointer to the `"]
-            #[doc = stringify!(#field_ident)]
-            #[doc = "` register."]
+            #[doc = "Obtain a pointer to the "]
+            #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+            #[doc = " register."]
             #[doc = ""]
             #[doc = "Never create a reference from this pointer - only use read/write/read_volatile/write_volatile methods on it."]
             #[inline(always)]
@@ -326,9 +367,9 @@ impl FieldParser {
                 unsafe { core::ptr::addr_of_mut!((*self.ptr).#field_ident) }
             }
 
-            #[doc = "Read the `"]
-            #[doc = stringify!(#field_ident)]
-            #[doc = "` register."]
+            #[doc = "Read the "]
+            #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+            #[doc = " register."]
             #[inline(always)]
             pub fn #read_fn_name(&mut self) -> #type_path {
                 let addr = self.#pointer_fn_name();
@@ -340,9 +381,9 @@ impl FieldParser {
 
         if access == Access::ReadWrite {
             access_methods.append_all(quote! {
-                #[doc = "Write the `"]
-                #[doc = stringify!(#field_ident)]
-                #[doc = "` register."]
+                #[doc = "Write the "]
+                #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+                #[doc = " register."]
                 #[inline(always)]
                 pub fn #write_fn_name(&mut self, value: #type_path) {
                     let addr = self.#pointer_fn_name();
@@ -351,9 +392,9 @@ impl FieldParser {
                     }
                 }
 
-                #[doc = "Read-Modify-Write the `"]
-                #[doc = stringify!(#field_ident)]
-                #[doc = "` register."]
+                #[doc = "Read-Modify-Write the "]
+                #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+                #[doc = " register."]
                 #[inline]
                 pub fn #modify_fn_name<F>(&mut self, f: F) where F: FnOnce(#type_path) -> #type_path {
                     let value = self. #read_fn_name();
@@ -365,6 +406,7 @@ impl FieldParser {
     }
     fn generate_array_access_methods(
         &self,
+        ident: &Ident,
         access: Access,
         field_ident: &Ident,
         type_array: &TypeArray,
@@ -382,9 +424,9 @@ impl FieldParser {
         let error_type = quote! { derive_mmio::OutOfBoundsError };
 
         access_methods.append_all(quote! {
-            #[doc = "Obtain a pointer to the `"]
-            #[doc = stringify!(#field_ident)]
-            #[doc = "` first entry register array."]
+            #[doc = "Obtain a pointer to the "]
+            #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+            #[doc = " first entry register array."]
             #[doc = ""]
             #[doc = "Never create a reference from this pointer - only use read/write/read_volatile/write_volatile methods on it."]
             #[doc = "The `add` method method of the pointer can be used to access entries of the array at higher indices."]
@@ -393,10 +435,9 @@ impl FieldParser {
                 unsafe { (*self.ptr).#field_ident.as_mut_ptr() }
             }
 
-            #[doc = "Read the `"]
-            #[doc = stringify!(#field_ident)]
-            #[doc = ""]
-            #[doc = "` register."]
+            #[doc = "Read the "]
+            #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+            #[doc = " register."]
             #[doc = ""]
             #[doc = "# Safety "]
             #[doc = ""]
@@ -411,10 +452,9 @@ impl FieldParser {
                 }
             }
 
-            #[doc = "Read the `"]
-            #[doc = stringify!(#field_ident)]
-            #[doc = ""]
-            #[doc = "` register."]
+            #[doc = "Read the "]
+            #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+            #[doc = " register."]
             #[doc = ""]
             #[doc = "This function also peforms bound checking."]
             #[inline]
@@ -432,9 +472,9 @@ impl FieldParser {
 
         if access == Access::ReadWrite {
             access_methods.append_all(quote! {
-                #[doc = "Write the `"]
-                #[doc = stringify!(#field_ident)]
-                #[doc = "` register."]
+                #[doc = "Write the "]
+                #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+                #[doc = " register."]
                 #[doc = "# Safety "]
                 #[doc = ""]
                 #[doc = "This function does not perform bounds checking and performs a volatile "]
@@ -448,9 +488,9 @@ impl FieldParser {
                     }
                 }
 
-                #[doc = "Write the `"]
-                #[doc = stringify!(#field_ident)]
-                #[doc = "` register."]
+                #[doc = "Write the "]
+                #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+                #[doc = " register."]
                 #[doc = ""]
                 #[doc = "This function also peforms bound checking."]
                 #[inline]
@@ -467,9 +507,9 @@ impl FieldParser {
                     Ok(())
                 }
 
-                #[doc = "Read-Modify-Write the `"]
-                #[doc = stringify!(#field_ident)]
-                #[doc = "` register."]
+                #[doc = "Read-Modify-Write the "]
+                #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+                #[doc = " register."]
                 #[doc = ""]
                 #[doc = "This function does not perform bounds checking and performs a volatile "]
                 #[doc = "read and a volatile write on a raw pointer with the given offset which might lead to "]
@@ -485,9 +525,9 @@ impl FieldParser {
                     self. #unchecked_write_fn_name(index, new_value);
                 }
 
-                #[doc = "Read-Modify-Write the `"]
-                #[doc = stringify!(#field_ident)]
-                #[doc = "` register."]
+                #[doc = "Read-Modify-Write the "]
+                #[doc = concat!("[", stringify!(#ident), "::", stringify!(#field_ident), "]")]
+                #[doc = " register."]
                 #[doc = ""]
                 #[doc = "This function also peforms bound checking."]
                 #[inline]